Rename: Renamed existing folders and files and added new folders and files, updated code paths due to folder restructuring
Renamed existing folders and updated file names as required.
Added new folders and files to the project directory.
Modified the code paths in the project to reflect the changes in folder structure and folder names.
