+++ app/components/Footer/Footer.module.css
@@ -1,0 +1,32 @@
+.footerContainer {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 100px;
+}
+
+.footerContents {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 8px;
+}
+
+.footerContents > p {
+  font-size: 20px;
+}
+
+.termsContainer {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+}
+
+.termsContainer > p {
+  font-size: 16px;
+}
+
+.termsContainer > p:hover {
+  cursor: pointer;
+}
